Rename update_financials.py to update_fitted_cost.py
