+++ smooth/components/component.py
@@ -1,5 +1,5 @@
 from oemof.outputlib import views
-from smooth.framework.functions.update_financials import update_financials, update_emissions
+from smooth.framework.functions.update_fitted_cost import update_financials, update_emissions
 from smooth.framework.functions.update_annuities import update_annuities
 
 class Component:
